js파일 병합 및 utils 생성
